% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbounds.R
\name{rdbounds_export}
\alias{rdbounds_export}
\title{Export Results from Manipulation Robust RD Estimation}
\usage{
rdbounds_export(rdbounds, file_name = NULL, view_it = FALSE)
}
\arguments{
\item{rdbounds}{An rdbounds object resulting from \code{\link{rdbounds}}. Required.}

\item{file_name}{base filename to output tables to. Expects a string of the form "path/filename", where filename has no extension and will be the root filename for a series of different files containing different tables. If omitted no files will be produced.}

\item{view_it}{Boolean. View main results table in Rstudio viewer. Defaults to \code{FALSE}.}
}
\description{
This function exports tables from manipulation robust RD estimation
}
\examples{
\donttest{df<-rdbounds_sampledata(50000, covs=TRUE)
rdbounds_est<-rdbounds(y=df$y,x=df$x, covs=as.factor(df$cov), treatment=df$treatment, c=0,
                       discrete_x=FALSE, discrete_y=FALSE,
                       bwsx=c(.2,.5), bwy = .1, kernel="epanechnikov", orders=1,
                       evaluation_ys = seq(from = 0, to=23, by=.2),
                       refinement_A=TRUE, refinement_B=TRUE,
                       right_effects=TRUE, yextremes = c(0,23),
                       num_bootstraps=0)
rdbounds_summary(rdbounds_est, title_prefix="Sample Data Results")}
}
