% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbounds.R
\name{rdbounds_sampledata}
\alias{rdbounds_sampledata}
\title{Generate a simulated dataset for testing estimation}
\usage{
rdbounds_sampledata(sample_size = 50000, covs = FALSE)
}
\arguments{
\item{sample_size}{Sample size for the dataset.}

\item{covs}{If set to \code{TRUE}, generates a sample in which half of the units have one of two covariate values, where the proportion of always-assigned units is slightly different for each.}
}
\description{
This function generates a simulated dataset with which to test \code{\link{rdbounds}}.
The x-values of potentially-assigned units (95% of sample) are normally distributed (variance of 5, censored at -10 and 10) around zero (the cutoff) and always-assigned units (5% of sample) follow a triangular distribution from the cutoff (x=0) to x=5. A random 5% of all units are never-takers, another 25% are always-takers, and the remaining units are compliers. Outcome values are then generated with a treatment effect of 2 for potentially-assigned units, 5 for always-assigned units, and following a linear trend with some normally distributed noise.
 Specifically: \deqn{y=(x+10)/2*treatment*(alwaysassigned=0)+5*treatment*(alwaysassigned=1)+normal(0,1))} and y is censored at 0 and 23.
}
\examples{
\donttest{df<-rdbounds_sampledata(50000);}
}
