% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdbounds.R
\name{rdbounds_summary}
\alias{rdbounds_summary}
\title{Summarize Results from Manipulation Robust RD Estimation}
\usage{
rdbounds_summary(rdbounds, title_prefix = "", text = TRUE)
}
\arguments{
\item{rdbounds}{an rdbounds object resulting from the function rdbounds(). Required.}

\item{title_prefix}{Optional prefix before "Average Treatment EFfects" or "Quantile Treatment Effects" in table.}

\item{text}{if set to \code{TRUE}, display results as text as well as formatted table. Defaults to \code{TRUE}.}
}
\description{
This function reports main estimands from \code{\link{rdbounds}}, as a formatted table, and optionally as text output.
}
\examples{
\donttest{df<-rdbounds_sampledata(50000, covs=TRUE)
rdbounds_est<-rdbounds(y=df$y,x=df$x, covs=as.factor(df$cov), treatment=df$treatment, c=0,
                       discrete_x=FALSE, discrete_y=FALSE, bwsx=c(.2,.5), bwy = .1,
                       kernel="epanechnikov", orders=1,
                       evaluation_ys = seq(from = 0, to=23, by=.2), ymin=0, ymax=23,
                       num_bootstraps=0, refinement_A=TRUE, refinement_B=TRUE, 
                       right_effects=TRUE, yextremes = c(0,23))
rdbounds_summary(rdbounds_est, title_prefix="Sample Data Results")}
}

